+++ src/commands/YCor.java
@@ -14,7 +14,7 @@ public class YCor extends AbstractZeroParameterTurtleCommand {
     @Override
     public int execute () {
         ITurtle turtle = getTurtle();
-        int position = (int) turtle.getY();
+        int position = turtle.getY();
         return position;
     }
